Allow extra $http config to be specified.
This is done by passing a fourth parameter to the $get, $post etc
methods of the resource. This parameter is an options map which will be
merged with the existing options map.
If an option named "httpConfig" is passed, the value of this will be
added to the $http config.

The purpose of this is to allow for custom http options to be specified
to specific requests.